feat: mock graphql query in storybook for search skill
