+++ packages/ui/src/components/Expandable/index.ts
@@ -1,0 +1,1 @@
+export * from "./Expandable";
