+++ packages/ui/src/components/SearchSkill/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchSkill";
