+++ packages/ui/src/components/index.ts
@@ -4,9 +4,10 @@ export * from "./BioComponent";
 export * from "./CollectionOfUsers";
 export * from "./Endorsements";
 export * from "./EndorsementsCarousel";
+export * from "./Expandable";
 export * from "./FormStepper";
 export * from "./LoginButton";
-export * from "./SearchBar";
+export * from "./SearchSkill";
 export * from "./SocialMediaComponent";
 export * from "./UserProfileMenu";
 export * from "./UserWithDescription";
